+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/commands.py
@@ -47,7 +47,8 @@ def load_command_table(self, _):  # pylint: disable=too-many-locals, too-many-st
         g.command('delete', 'delete', confirmation=True)
         g.show_command('show', 'get_properties')
         g.custom_command('list', 'list_storage_accounts')
-        g.custom_command('show-usage', 'show_storage_account_usage')
+        g.custom_command('show-usage', 'show_storage_account_usage', min_api='2018-03-01-preview')
+        g.custom_command('show-usage', 'show_storage_account_usage_no_location', max_api='2016-01-01')
         g.custom_command('show-connection-string', 'show_storage_account_connection_string')
         g.generic_update_command('update', getter_name='get_properties', setter_name='update',
                                  custom_func_name='update_storage_account', min_api='2016-12-01')
@@ -223,7 +224,7 @@ def load_command_table(self, _):  # pylint: disable=too-many-locals, too-many-st
 
         g.storage_custom_command('policy create', 'create_acl_policy')
         g.storage_custom_command('policy delete', 'delete_acl_policy')
-        g.storage_custom_command('policy show', 'get_acl_policy')
+        g.storage_custom_command('policy show', 'get_acl_policy', exception_handler=show_exception_handler)
         g.storage_custom_command('policy list', 'list_acl_policies', table_transformer=transform_acl_list_output)
         g.storage_custom_command('policy update', 'set_acl_policy')