+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+2.2.2
+++++++
+* `vm/vmss identity show`: exception handling to exit with code 3 upon a missing resource for consistency
+* `vm create`: deprecate `--storage-caching` option.
+
 2.2.1
 ++++++
 * vm list-skus: add a few common filters to make the command easier to use